backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
langgraph==0.0.26
langchain==0.1.0
langchain-core==0.1.10
langchain-openai==0.0.2
